simplified page module structure
