+++ spec/features/members/error_messages_spec.rb
@@ -36,7 +36,7 @@ feature 'group memberships through groups page', type: :feature do
 
   let!(:manager) { FactoryGirl.create :role, name: 'Manager' }
 
-  let(:members_page) { Pages::Project::Members.new project.identifier }
+  let(:members_page) { Pages::Members.new project.identifier }
 
   before do
     allow(User).to receive(:current).and_return admin
